@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\ViewModel;
 
-use App\Entity\TrackUpload;
+use App\Entity\Track;
 use Ramsey\Uuid\UuidInterface;
 
 final class TrackToReview
@@ -14,17 +14,15 @@ final class TrackToReview
     private UuidInterface $id;
     private string $filename;
     private ?string $title;
-    private ?array $artists;
     private ?array $albumArtists;
     private ?string $album;
     private ?int $trackNo;
-    private $status;
+    private string $status;
 
     public function __construct(
         UuidInterface $id,
         string $filename,
         ?string $title,
-        ?array $artists,
         ?array $albumArtists,
         ?string $album,
         ?int $trackNo
@@ -32,7 +30,6 @@ final class TrackToReview
         $this->id = $id;
         $this->filename = $filename;
         $this->title = $title;
-        $this->artists = $artists;
         $this->albumArtists = $albumArtists;
         $this->album = $album;
         $this->trackNo = $trackNo;
@@ -47,11 +44,6 @@ final class TrackToReview
     public function getFilename(): string
     {
         return $this->filename;
-    }
-
-    public function getArtists(): ?array
-    {
-        return $this->artists;
     }
 
     public function getAlbumArtists(): ?array
@@ -74,16 +66,15 @@ final class TrackToReview
         return $this->status;
     }
 
-    public static function fromUpload(TrackUpload $upload): TrackToReview
+    public static function fromUpload(Track $upload): TrackToReview
     {
         return new self(
             $upload->getId(),
-            $upload->getFilename(),
-            $upload->getMetadata()->title,
-            $upload->getMetadata()->artists,
-            $upload->getMetadata()->albumArtists,
-            $upload->getMetadata()->album,
-            $upload->getMetadata()->trackNo
+            $upload->getTitle(),
+            $upload->getTitle(), // @fixme
+            $upload->getAlbum() !== null ? $upload->getAlbum()->getArtists() : [],
+            $upload->getAlbum() !== null ? $upload->getAlbum()->getTitle() : null,
+            $upload->getTrackNo()
         );
     }
 }
